# Conflicts:
#   service/transporter/controller/processor.go

+++ service/httpserver/controller/fileController.go
@@ -7,6 +7,7 @@ import (
 	"cloud-storage-httpserver/service/scheduler"
 	"cloud-storage-httpserver/service/tools"
 	"cloud-storage-httpserver/service/transporter"
+	"fmt"
 	"net/http"
 	"strings"
 
@@ -18,12 +19,12 @@ func UserGetFiles(con *gin.Context) {
 		args.FieldWordAccessToken: true,
 		args.FieldWordFilePath:    true,
 	}
-	valueMap, existMap := getQueryAndReturn(con, fieldRequired)
-	if tools.RequiredFieldNotExist(fieldRequired, existMap) {
+	valueMap, existMap := getQueryAndReturn(con, &fieldRequired)
+	if tools.RequiredFieldNotExist(&fieldRequired, existMap) {
 		return
 	}
-	accessToken := valueMap[args.FieldWordAccessToken].(string)
-	filePath := valueMap[args.FieldWordFilePath].(string)
+	accessToken := (*valueMap)[args.FieldWordAccessToken].(string)
+	filePath := (*valueMap)[args.FieldWordFilePath].(string)
 	//check token
 	userId, valid := UserCheckAccessToken(con, accessToken)
 	if !valid {
@@ -49,7 +50,7 @@ func UserGetFiles(con *gin.Context) {
 	//	})
 	//	return
 	//}
-	var integrateFiles []model.FileAndDir
+	var integrateFiles []model.FileAndDir = make([]model.FileAndDir, 0)
 	dirMap := make(map[string]model.File)
 	for _, file := range *files {
 		remainString := strings.TrimPrefix(file.FileName, filePath)
@@ -113,12 +114,12 @@ func UserPreUploadFile(con *gin.Context) {
 		args.FieldWordAccessToken: true,
 		args.FieldWordFilePath:    true,
 	}
-	valueMap, existMap := getQueryAndReturn(con, fieldRequired)
-	if tools.RequiredFieldNotExist(fieldRequired, existMap) {
+	valueMap, existMap := getQueryAndReturn(con, &fieldRequired)
+	if tools.RequiredFieldNotExist(&fieldRequired, existMap) {
 		return
 	}
-	accessToken := valueMap[args.FieldWordAccessToken].(string)
-	filePath := valueMap[args.FieldWordFilePath].(string)
+	accessToken := (*valueMap)[args.FieldWordAccessToken].(string)
+	filePath := (*valueMap)[args.FieldWordFilePath].(string)
 	//check access token
 	userId, valid := UserCheckAccessToken(con, accessToken)
 	if !valid {
@@ -143,6 +144,16 @@ func UserPreUploadFile(con *gin.Context) {
 		})
 		return
 	}
+	// check user status
+	statusMap := map[string]bool{
+		args.UserForbiddenStatus: false,
+		args.UserVerifyStatus:    false,
+	}
+
+	if !UserCheckStatus(con, user, &statusMap) {
+		return
+	}
+
 	// preUpload and get a token
 	response, success := transporter.PreUploadFile(filePath, user)
 	if !success {
@@ -155,6 +166,9 @@ func UserPreUploadFile(con *gin.Context) {
 	}
 	// error in transporter
 	if response.Code != args.CodeOK {
+		fmt.Println("transporter fault:")
+		fmt.Println("Code: ", response.Code)
+		fmt.Println("Msg: ", response.Msg)
 		con.JSON(http.StatusOK, gin.H{
 			"code": response.Code,
 			"msg":  response.Msg,
@@ -177,12 +191,12 @@ func UserDownloadFile(con *gin.Context) {
 		args.FieldWordAccessToken: true,
 		args.FieldWordFilePath:    true,
 	}
-	valueMap, existMap := getQueryAndReturn(con, fieldRequired)
-	if tools.RequiredFieldNotExist(fieldRequired, existMap) {
+	valueMap, existMap := getQueryAndReturn(con, &fieldRequired)
+	if tools.RequiredFieldNotExist(&fieldRequired, existMap) {
 		return
 	}
-	accessToken := valueMap[args.FieldWordAccessToken].(string)
-	filePath := valueMap[args.FieldWordFilePath].(string)
+	accessToken := (*valueMap)[args.FieldWordAccessToken].(string)
+	filePath := (*valueMap)[args.FieldWordFilePath].(string)
 	userId, valid := UserCheckAccessToken(con, accessToken)
 	if !valid {
 		return
@@ -213,7 +227,44 @@ func UserDownloadFile(con *gin.Context) {
 			"data": gin.H{},
 		})
 	}
-	// use scheduler's download plan to download file with transporter
+	// check user status
+	statusMap := map[string]bool{
+		args.UserForbiddenStatus: false,
+		args.UserVerifyStatus:    false,
+	}
+	if !UserCheckStatus(con, user, &statusMap) {
+		return
+	}
+	// check file status if done -> return url
+	files, success := dao.FileDao.CheckFileStatus(userId, filePath)
+	if !success {
+		con.JSON(http.StatusOK, gin.H{
+			"code": args.CodeDatabaseError,
+			"msg":  "数据库错误",
+			"data": gin.H{},
+		})
+		return
+	}
+	var doneURL string
+	var doneFlag bool = false
+	for _, file := range *files {
+		if file.ReconstructStatus == args.FileReconstructStatusDone {
+			doneFlag = true
+			doneURL = file.DownloadUrl
+		}
+	}
+	if doneFlag {
+		con.JSON(http.StatusOK, gin.H{
+			"code": args.CodeOK,
+			"msg":  "解析scheduler-json信息有误",
+			"data": gin.H{
+				"Type":   "url",
+				"Result": doneURL,
+			},
+		})
+		return
+	}
+	// else -> use scheduler's download plan to download file with transporter
 	getDownloadPlanResponse, success := scheduler.GetDownloadPlanFromScheduler(userId, filePath)
 	if !success {
 		con.JSON(http.StatusOK, gin.H{
@@ -234,6 +285,9 @@ func UserDownloadFile(con *gin.Context) {
 	}
 	// error in transporter
 	if downloadResponse.Code != args.CodeOK {
+		fmt.Println("transporter fault:")
+		fmt.Println("Code: ", downloadResponse.Code)
+		fmt.Println("Msg: ", downloadResponse.Msg)
 		con.JSON(http.StatusOK, gin.H{
 			"code": downloadResponse.Code,
 			"msg":  downloadResponse.Msg,
@@ -254,12 +308,12 @@ func UserDeleteFile(con *gin.Context) {
 		args.FieldWordAccessToken: true,
 		args.FieldWordFilePath:    true,
 	}
-	valueMap, existMap := getQueryAndReturn(con, fieldRequired)
-	if tools.RequiredFieldNotExist(fieldRequired, existMap) {
+	valueMap, existMap := getQueryAndReturn(con, &fieldRequired)
+	if tools.RequiredFieldNotExist(&fieldRequired, existMap) {
 		return
 	}
-	accessToken := valueMap[args.FieldWordAccessToken].(string)
-	filePath := valueMap[args.FieldWordFilePath].(string)
+	accessToken := (*valueMap)[args.FieldWordAccessToken].(string)
+	filePath := (*valueMap)[args.FieldWordFilePath].(string)
 	// check token
 	userId, valid := UserCheckAccessToken(con, accessToken)
 	if !valid {
@@ -284,6 +338,15 @@ func UserDeleteFile(con *gin.Context) {
 		})
 		return
 	}
+	// check user status
+	statusMap := map[string]bool{
+		args.UserForbiddenStatus: false,
+		args.UserVerifyStatus:    false,
+	}
+	if !UserCheckStatus(con, user, &statusMap) {
+		return
+	}
+
 	// delete file with transporter
 	success = transporter.DeleteFile(filePath, user)
 	if !success {
